requirements.txt
anyio==3.6.2
ask-sdk-core==1.17.1
ask-sdk-model==1.35.0
ask-sdk-runtime==1.17.1
ask-sdk-webservice-support==1.3.3
asn1crypto==1.5.1
async-timeout==4.0.2
Authlib==1.1.0
black==22.10.0
cachetools==5.2.0
certifi==2022.9.24
certvalidator==0.11.1
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cryptography==38.0.1
Deprecated==1.2.13
dnspython==2.2.1
ecdsa==0.18.0
fastapi==0.85.1
freezegun==1.2.2
google-auth==2.14.0
google-auth-oauthlib==0.7.0
h11==0.12.0
httpcore==0.15.0
httptools==0.5.0
httpx==0.23.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy-extensions==0.4.3
oauthlib==3.2.2
oscrypto==1.3.0
packaging==21.3
pathspec==0.10.1
platformdirs==2.5.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.2
pymongo==4.3.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
python-jose==3.3.0
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
ssml-builder==1.1.1
starlette==0.20.4
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.18.3
watchfiles==0.18.0
websockets==10.3
wrapt==1.14.1
